Move `ShareAnnotationsPanel` => `ShareDialog`
